+++ client/src/styles/app.scss
@@ -4,4 +4,3 @@
 @import './header.scss';
 @import './search.scss';
 @import './cards.scss';
-
