+++ client/src/styles/login.scss
@@ -1,13 +1,3 @@
-#login-input {
-    border: none;
-    box-shadow: 0 5px 10px 0 rgba(0, 0, 0, .2);
-    margin-bottom: 2rem;
-    width: 25vw;
-    height: 50px;
-    font-size: 24px;
-    padding-left: 8px;
-    background: white;
-}
 .login-screen {
     height: 100vh;
     display: flex;
@@ -20,6 +10,7 @@
     background-repeat: no-repeat;
     background-position: 20% 45%;
 }
+
 .login-wrapper {
     display: flex;
     flex-direction: column;
@@ -39,42 +30,76 @@
         text-align: center;
         margin-bottom: 5%;
     }
+
+    .error-modal {
+        background: #ea1601;
+        color: white;
+        border-radius: 10px;
+        width: 55%;
+        padding: 2%;
+    }
+
+    .success-modal {
+        background: #64bc26;
+        color: white;
+        border-radius: 10px;
+        width: 55%;
+        padding: 2%;
+    }
+    
+    .errors {
+        color: red;
+    }
     form {
         display: flex;
         flex-direction: column;
         align-items: center;
         width: 100%;
-        .MuiFormControl-root {
+
+
+        .login-input {
             margin-bottom: 5%;
-            width: 358px;
             background: #ffffff;
-            width: 60%;
+            width: 65%;
             box-shadow: 0px 7px 29px rgba(0, 0, 0, 0.25);
-            
-            input {
-                font-size: 2.5rem;
+            font-size: 2.4rem;
+            border: none;
+            height: 50px;
+            padding-left: 3%;
+        }
+
+        .login-btn {
+            background: #FCDE7B;
+            border: none;
+            font-size: 2.8rem;
+            width: 65%;
+            margin-top: 2%;
+            margin-bottom: 5%;
+            box-shadow: 0px 7px 29px rgba(0, 0, 0, 0.25);
+            text-transform: none;
+            color: black;
+            transition: .25s ease-in-out;
+            cursor: pointer;
+
+            &:hover {
+                background: coral;
+                color: white;
             }
         }
-    }
-    .options, a {
-        font-size: 2rem;
-        color: #F17012;
-    }
-}
-.cookin-button-login {
-    background: #FCDE7B;
-    font-size: 2.8rem;
-    box-shadow: 0px 7px 29px rgba(0, 0, 0, 0.25);
-    border: none;
-    transition: .25s ease-in-out;
-    align-self: center;
-    margin-top: 3rem;
-    margin-bottom: 3rem;
-    width: 25vw;
 
-    &:hover {
-        background: coral;
-        color: white;
-        cursor: pointer;
+        .options {
+            font-size: 2rem;
+
+            a {
+                color: #f17012;
+
+                &:hover {
+                    border-bottom: 2px solid #f17012;
+                }
+                
+            }
+        }
+
+        
     }
 }
