--- frontend/client/src/styles/app.scss
@@ -1,6 +1,0 @@
-@import './login.scss';
-@import './signup.scss';
-@import './landing.scss';
-@import './header.scss';
-@import './search.scss';
-@import './cards.scss';
