--- frontend/client/src/styles/header.scss
@@ -1,26 +1,0 @@
-.header {
-    width: 100vw;
-    background: linear-gradient(180deg, #FFFFFF 0%, rgba(255, 255, 255, 0) 84.37%), #FCDE7B;
-    box-shadow: 0px 7px 29px rgba(0, 0, 0, 0.25);
-    position: fixed;
-    top: 0;
-    height: 60px;
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-}
-.header-content {
-    width: 12%;
-    display: flex;
-    justify-content: space-evenly;
-}
-#liked-recipes-button {
-    border-radius: 50px;
-    background: transparent;
-    transition: .25s ease-in-out;
-
-    &:hover {
-        background: coral;
-        color: white;
-    }
-}
