--- frontend/client/src/styles/landing.scss
@@ -1,98 +1,0 @@
-.content {
-    background: url('../assets/reciplease-landing.png');
-    height: 100vh;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    padding: 0 2rem;
-    background-size: cover;
-}
-.landing-title {
-    font-family: 'Rock Salt';
-    font-size: 72px;
-    margin-bottom: 7rem;
-}
-.landing-description {
-    font-family: Roboto;
-    font-size: 28px;
-    width: 30%;
-    text-align: center;
-    margin-bottom: 7rem;
-    font-weight: 400;
-}
-
-.loading_container {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-.card_container {
-    margin-top: 3vh;
-    width: 84vw;
-    height: 40vh;
-    display: flex;
-    align-items: center;
-    flex-wrap: nowrap;
-    overflow-x: scroll;
-
-    .card {
-        flex: 0 0 auto;
-        width: 250px;
-        height: 90%;
-        margin: 0 1.5vw 0;
-        background-color: white;
-        overflow-y: scroll;
-
-        &::-webkit-scrollbar {
-            display: none;
-        }
-
-        img {
-            width: 100%;
-        }
-
-        .card_content {
-            padding: 10px;
-
-            h1 {
-                font-size: 1.8rem;
-                text-align: center;
-                font-weight: 500;
-            }
-    
-            p {
-                font-size: 1.2rem;
-                margin-top: 5px;
-                line-height: 1.5;
-            }
-        }
-
-    }
-}
-.not_active {
-    &::-webkit-scrollbar {
-        display: none;
-    }
-}
-#landing-btn {
-    font-family: Roboto;
-    width: 30%;
-    background: #fcde7b;
-    border: none;
-    font-size: 24px;
-    font-weight: 400;
-    box-shadow: 0 5px 10px 0 rgba(0, 0, 0, .15);
-    transition: .25s ease-in-out;
-    padding: 2rem 2rem;
-    border-radius: .5rem;
-
-    &:hover {
-        background: coral;
-        cursor: pointer;
-        color: white;
-        box-shadow: 0 10px 20px 0 rgba(0, 0, 0, .3);
-    }
-}
