--- frontend/client/src/styles/login.scss
@@ -1,105 +1,0 @@
-.login-screen {
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    background: url('../assets/edgar-castrejon-1CsaVdwfIew-unsplash.jpg');
-    background-size: cover;
-    background-repeat: no-repeat;
-    background-position: 20% 45%;
-}
-
-.login-wrapper {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    width: 40%;
-    background: rgba(235, 244, 255, 0.95);
-    h1 {
-        font-size: 6.4rem;
-        margin-top: 5%;
-        margin-bottom: 4%;
-        font-weight: bold;
-    }
-    p {
-        font-size: 1.8rem;
-        width: 454px;
-        text-align: center;
-        margin-bottom: 5%;
-    }
-
-    .error-modal {
-        background: #ea1601;
-        color: white;
-        border-radius: 10px;
-        width: 55%;
-        padding: 2%;
-    }
-
-    .success-modal {
-        background: #64bc26;
-        color: white;
-        border-radius: 10px;
-        width: 55%;
-        padding: 2%;
-    }
-    
-    .errors {
-        color: red;
-    }
-    form {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        width: 100%;
-
-
-        .login-input {
-            margin-bottom: 5%;
-            background: #ffffff;
-            width: 65%;
-            box-shadow: 0px 7px 29px rgba(0, 0, 0, 0.25);
-            font-size: 2.4rem;
-            border: none;
-            height: 50px;
-            padding-left: 3%;
-        }
-
-        .login-btn {
-            background: #FCDE7B;
-            border: none;
-            font-size: 2.8rem;
-            width: 65%;
-            margin-top: 2%;
-            margin-bottom: 5%;
-            box-shadow: 0px 7px 29px rgba(0, 0, 0, 0.25);
-            text-transform: none;
-            color: black;
-            transition: .25s ease-in-out;
-            cursor: pointer;
-
-            &:hover {
-                background: coral;
-                color: white;
-            }
-        }
-
-        .options {
-            font-size: 2rem;
-
-            a {
-                color: #f17012;
-
-                &:hover {
-                    border-bottom: 2px solid #f17012;
-                }
-                
-            }
-        }
-
-        
-    }
-}
