--- frontend/client/src/styles/search.scss
@@ -1,93 +1,0 @@
-.search-content {
-    height: 100vh;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 12rem 2rem 4rem 2rem;
-    background: url('../assets/reciplease-search.png');
-    background-size: cover;
-
-    .search-form {
-        display: flex;
-        width: 90rem;
-        transition: opacity .75s;
-        opacity: 1;
-        
-        .search-input {
-            background: linear-gradient(180deg, rgba(255, 255, 255, 0) 80.21%, rgba(0, 0, 0, 0.25) 100%),
-            linear-gradient(0deg, #FFFFFF, #FFFFFF);
-            height: 5.5rem;
-            width: 60rem;
-            margin-left: 6rem;
-            margin-right: .3rem;
-            line-height: 87%;
-            font-size: 2.4rem;
-            padding-left: 3%;
-        }
-
-        .search-button {
-            background: linear-gradient(180deg, rgba(255, 255, 255, 0) 88.02%, #B8B6B6 100%),
-            linear-gradient(180deg, #FFFFFF 0%, rgba(255, 255, 255, 0) 55.73%),
-            linear-gradient(0deg, #FCDE7B, #FCDE7B);
-            font-family: Roboto;
-            font-size: 3rem;
-            font-weight: 400;
-            text-align: center;
-            width: 182px;
-            height: 5.5rem;
-            transition: .25s ease-in-out;
-        
-            &:hover {
-                cursor: pointer;
-                color: white;
-                background: coral;
-            }
-        }
-    }
-}
-
-.drawer-container {
-    display: flex;
-    flex-direction: column;
-    .drawer-recipe-image {
-        width: 500px;
-        display: flex;
-        align-self: center;
-        justify-self: center;
-
-        object-fit: cover;
-        border-radius: 10px;
-        margin-top: 4%;
-    }
-
-    .drawer-card-content {
-        width: 100%;
-        padding-left: 2rem;
-        padding-right: 1rem;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        border: none;
-        padding-top: 2%;
-
-        .drawer-recipe-title {
-            font-size: 2rem;
-            width: 70%;
-            text-align: center;
-        }
-
-        .drawer-recipe-info {
-            display: flex;
-            justify-content: space-between;
-            width: 75%;
-            margin-top: 6%;
-
-            .description {
-                font-size: 2rem;
-                text-align: center;
-                line-height: 1.5;
-            }
-        }
-    }
-}
