--- frontend/client/src/styles/signup.scss
@@ -1,98 +1,0 @@
-.signup-container {
-    height: 100vh;
-    width: 100vw;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    background: url("../assets/signup-page.png");
-    background-size: cover;
-
-    form {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        font-size: 3rem;
-        background: rgba(235, 244, 255, 0.95);
-        width: 40%;
-
-        h3 {
-            text-align: center;
-            margin-bottom: 3%;
-            font-size: 6.4rem;
-            font-weight: bold;
-            margin-top: 5%;
-        }
-
-        p {
-            font-size: 1.8rem;
-            text-align: center;
-            margin-bottom: 4%;
-            width: 454px;
-            line-height: 2.4rem;
-        }
-
-        .errors {
-            color: red;
-        }
-
-        .error-modal {
-            background: #ea1601;
-            color: white;
-            border-radius: 10px;
-            width: 55%;
-            padding: 2%;
-        }
-
-        .success-modal {
-            background: #64bc26;
-            color: white;
-            border-radius: 10px;
-            width: 55%;
-            padding: 2%;
-        }
-
-        .signup-input {
-            margin-bottom: 5%;
-            background: #ffffff;
-            width: 65%;
-            box-shadow: 0px 7px 29px rgba(0, 0, 0, 0.25);
-            font-size: 2.4rem;
-            border: none;
-            height: 50px;
-            padding-left: 3%;
-        }
-
-        .signup-btn {
-            background: #FCDE7B;
-            border: none;
-            font-size: 2.8rem;
-            width: 65%;
-            margin-top: 2%;
-            margin-bottom: 5%;
-            box-shadow: 0px 7px 29px rgba(0, 0, 0, 0.25);
-            text-transform: none;
-            color: black;
-            transition: .25s ease-in-out;
-            cursor: pointer;
-
-            &:hover {
-                background: coral;
-                color: white;
-            }
-        }
-
-        .options {
-            font-size: 2rem;
-
-            a {
-                color: #f17012;
-
-                &:hover {
-                    border-bottom: 2px solid #f17012;
-                }
-                
-            }
-        }
-    }
-}
